Componentnt and Object fixes 1/2

+++ Engine/Core/Timer.cpp
@@ -25,7 +25,7 @@ namespace nc {
 	}
 
 	void FrameTimer::Tick() {
-		m_dt = ElapsedSeconds();
+		m_dt = static_cast<float>(ElapsedSeconds());
 		m_tp = clock::now();
 	}
 }
